+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changes
 =======
 
+Unreleased
+----------
+
+* ControlMessage and ControlRequest are now simply Message and Request,
+  as there is no other kind of message or request.
+
 0.0.3 (2019-05-03)
 ------------------